pyproject.toml
[project]
name = "webflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "2captcha-python>=2.0.3",
    "pydantic>=2.12.5",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.2",
]
